+++ R/add_heuristic_solver.R
@@ -25,6 +25,10 @@ NULL
 #'  \item All actions are initially selected for funding (excepting actions
 #'    which are locked out).
 #'
+#'  \item Actions which exceed the budget are then deselected. Actions which
+#'    are exclusively associated with projects that exceed the budget
+#'    are also deselected.
+#'
 #'  \item The next step is repeated until (i) the number of desired
 #'    solutions is obtained and (ii) the total cost of the remaining
 #'    actions that are selected for funding is within the budget.
@@ -57,6 +61,10 @@ NULL
 #'
 #'  \item All actions are initially selected for funding (excepting actions
 #'    which are locked out).
+#'
+#'  \item Actions which exceed the budget are then deselected. Actions which
+#'    are exclusively associated with projects that exceed the budget
+#'    are also deselected.
 #'
 #'  \item The next step is repeated until (i) the number of desired
 #'    solutions is obtained or (ii) no action can be deselected for funding
@@ -166,7 +174,7 @@ add_heuristic_solver <- function(x, number_solutions = 1, verbose = TRUE) {
         targets <- rep(0, length(fp$tip.label))
       }
       if (inherits(x$data$objective, "MinimumSetObjective")) {
-        budget <- -1
+        budget <- Inf
       } else {
         budget <- x$data$objective$parameters$get("budget")
       }